@@ -4,17 +4,17 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"os"
 )
 
 func Fixture(fname string) io.ReadCloser {
 
 	fpath := fmt.Sprintf("testdata/%s", fname)
 
-	f, err := ioutil.ReadFile(fpath)
+	f, err := os.ReadFile(fpath)
 	if err != nil {
 		panic(fmt.Sprintf("Cannot read fixture file %s: %s", fpath, err))
 	}
 
-	return ioutil.NopCloser(bytes.NewReader(f))
+	return io.NopCloser(bytes.NewReader(f))
 }
